pyproject.toml
[project]
name = "ai-web-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "aiosqlite>=0.20.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "pillow>=11.0.0",
    "bcrypt==4.0.1",
    "python-docx>=1.2.0",
    "markdown>=3.6",
    "markdown2>=2.5.3",
    "beautifulsoup4>=4.13.4",
    "fpdf2>=2.7.8",
    "pypdf>=5.0.0",
]
